Move CutText into Core/Text text utilities
CutText is a pure wide-string split helper, not chat-specific, so it belongs
with SeparateTextIntoLines in Core/Text/TextLineWrap rather than ZzzInterface.cpp.
Callers (Notices, Chat, PersonalShopTitle) now include TextLineWrap.h. No
behaviour change.
