+++ src/source/Core/Text/TextLineWrap.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef> // size_t
+
 // Word-wrap utilities for wide-character UI text.
 // Pure function unit — no project dependencies, can be linked from tests.
 
@@ -12,3 +14,12 @@
  * @return int          Total lines created (guaranteed not to exceed iMaxLine).
  */
 int SeparateTextIntoLines(const wchar_t* lpszText, wchar_t* lpszSeparated, int iMaxLine, int iLineSize);
+
+/**
+ * @brief Splits Text into two parts near its midpoint at a space.
+ * @param Text       Source string.
+ * @param Text1      Receives the first part (up to maxLength, truncated).
+ * @param Text2      Receives the second part, or empty if no space was found.
+ * @param maxLength  Capacity of each destination buffer.
+ */
+void CutText(const wchar_t* Text, wchar_t* Text1, wchar_t* Text2, size_t maxLength);
